requirements.txt
#python 3.7
django == 3.0.5
djangorestframework == 3.11.0
markdown == 3.2.1
django-filter == 2.2.0